Split up serialization code into smaller moduler
