+++ serde_arrow/src/internal/serialization/interpreter.rs
@@ -11,17 +11,17 @@ use crate::{
         error::{self, fail, Result},
         serialization::{
             bit_set::BitSet,
-            compiler::{
-                dispatch_bytecode, BufferCounts, Bytecode, LargeListEnd, LargeListItem,
-                LargeListStart, ListEnd, ListItem, ListStart, MapEnd, MapItem, MapStart,
-                OptionMarker, OuterRecordEnd, OuterRecordField, OuterRecordStart, OuterSequenceEnd,
-                OuterSequenceItem, OuterSequenceStart, Panic, Program, ProgramEnd, PushBool,
-                PushDate64FromNaiveStr, PushDate64FromUtcStr, PushDictionary, PushF16, PushF32,
-                PushF64, PushI16, PushI32, PushI64, PushI8, PushLargeUtf8, PushNull, PushU16,
-                PushU32, PushU64, PushU8, PushUtf8, StructEnd, StructField, StructItem,
-                StructStart, Structure, TupleStructEnd, TupleStructItem, TupleStructStart,
-                UnionEnd, Variant,
+            bytecode::{
+                dispatch_bytecode, Bytecode, LargeListEnd, LargeListItem, LargeListStart, ListEnd,
+                ListItem, ListStart, MapEnd, MapItem, MapStart, OptionMarker, OuterRecordEnd,
+                OuterRecordField, OuterRecordStart, OuterSequenceEnd, OuterSequenceItem,
+                OuterSequenceStart, Panic, ProgramEnd, PushBool, PushDate64FromNaiveStr,
+                PushDate64FromUtcStr, PushDictionary, PushF16, PushF32, PushF64, PushI16, PushI32,
+                PushI64, PushI8, PushLargeUtf8, PushNull, PushU16, PushU32, PushU64, PushU8,
+                PushUtf8, StructEnd, StructField, StructItem, StructStart, TupleStructEnd,
+                TupleStructItem, TupleStructStart, UnionEnd, Variant,
             },
+            compiler::{BufferCounts, Program, Structure},
         },
         sink::EventSink,
     },
