+++ serde_arrow/src/internal/serialization/mod.rs
@@ -1,6 +1,8 @@
 pub mod bit_set;
+pub mod bytecode;
 pub mod compiler;
 pub mod interpreter;
+pub mod structure;
 
 pub use compiler::{compile_serialization, CompilationOptions};
 pub use interpreter::Interpreter;
